+++ app/Plugin/Schedules_Base.php
@@ -122,11 +122,11 @@ class Schedules_Base {
 		}
 
 		if ( ! wp_next_scheduled( $this->get_name() ) ) {
-			// create the schedule
+			// create the schedule.
 			$result = wp_schedule_event( time(), $this->get_interval(), $this->get_name(), $this->get_args() );
 
 			// log event if the schedule could not be created.
-			if( is_wp_error( $result ) ) {
+			if ( is_wp_error( $result ) ) {
 				Log::get_instance()->add( __( 'Error during creation of schedule:', 'personio-integration-light' ) . ' <code>' . wp_json_encode( $result->get_error_message() ) . '</code>', 'error', $this->get_log_category() );
 			}
 		}
@@ -142,7 +142,7 @@ class Schedules_Base {
 		$result = wp_clear_scheduled_hook( $this->get_name(), $this->get_args() );
 
 		// log event if the schedule could not be deleted.
-		if( is_wp_error( $result ) ) {
+		if ( is_wp_error( $result ) ) {
 			Log::get_instance()->add( __( 'Error during deleting of schedule:', 'personio-integration-light' ) . ' <code>' . wp_json_encode( $result->get_error_message() ) . '</code>', 'error', $this->get_log_category() );
 		}
 	}