+++ changelog.md
@@ -7,6 +7,11 @@
 - Text optimization in the import log
 - Optimization of the last PHPStan warnings
 - Optimization of schedule handling
+- Updated dependencies
+
+### Fixed
+
+- Fixed group of positions which were not correct sorted
 
 ## [5.0.1] - 27.10.2025