+++ changelog.md
@@ -12,6 +12,7 @@
 ### Fixed
 
 - Fixed group of positions which were not correct sorted
+- Fixed wrong email headers
 
 ## [5.0.1] - 27.10.2025